Move lineage api type definitions to api/
